Use CSS to style separators in comments info
As mentioned in commit 186164be0, ERB Lint is reporting some false
positives about hard coded translations due to these separators.

Note we're positioning the separators using the `::before`
pseudoelement. When using the `::after` pseudoelement, Firefox was
introducing an additional blank line.

For consistency with other mixins, we're making the `::before`
pseudoelement the default in the `circle-separators` mixin and using
`::after` in the places where `::before` would overwrite icons.