Use CSS to style separators in comment actions
As mentioned in commit 186164be0, ERB Lint is reporting some false
positives about hard coded translations due to these separators.

Now that we don't have to pass the `divider` in the comments controller,
we can reuse the `FlagActions` concern.

Since we're no longer using a `divider` HTML class, we're removing all
CSS references to that class.