Extract component to render comment reply button and actions
We were using the `reply` HTML class. This name was misleading because
this element doesn't show a reply, but it shows the responses count, the
"reply" form, the votes, the possibility to flag it or delete it, ... It
was even more misleading inside a notification, because there we only
render moderation actions.

So we're renaming it to `StatusBar`.

To simplify the code, we're using `current_user` instead of
`user_signed_in?`, since we use `current_user` for often.