Use currency in translations in investment dossiers
This is the same we do for the form to edit the investment. This way we
don't use parenthesis even for languages that wouldn't use parenthesis
in this case.

Note we're changing the i18n key for these texts. The reason is that
translations for existing languages are obsolete and, if we keep the
same translation key, existing languages will no longer display the
currency until translators update them.

Also note that, in the `_dossier_detail` partial, we were passing the
currency to the `show.price_first_year` translation, but we weren't
using the currency in that translation, and so it wasn't displayed.
We're fixing that as well.