Add HardCodedString ERB Lint rule
Note we aren't checking the proposal notification digest view; it's got
a false positive and replacing it with CSS is hard due to it being a
mailer view.

We're also excluding the `.js.erb` files. ERB Lint already excludes them
by default, but Pronto was checking them and reporting false positives.